Updated uTinyRipper Upstream Improved formating of dump info Improved formatting of shader output

+++ Extract/Util.cs
@@ -7,10 +7,12 @@ using System.Security.Cryptography;
 using System.Text;
 using System.Threading;
 using uTinyRipper;
-using uTinyRipper.Assembly;
 using uTinyRipper.Classes;
-using uTinyRipper.SerializedFiles;
+using uTinyRipper.Classes.Misc;
+using uTinyRipper.Game.Assembly;
+using uTinyRipper.Layout;
 using YamlDotNet.RepresentationModel;
+using Version = uTinyRipper.Version;
 
 namespace Extract
 {
@@ -55,6 +57,19 @@ namespace Extract
 				DeleteDir(dir);
 			}
 		}
+		public static string HashBytes(byte[] inputBytes)
+		{
+			using (MD5 md5 = MD5.Create())
+			{
+				byte[] hashBytes = md5.ComputeHash(inputBytes);
+				StringBuilder sb = new StringBuilder();
+				for (int i = 0; i < hashBytes.Length; i++)
+				{
+					sb.Append(hashBytes[i].ToString("X2"));
+				}
+				return sb.ToString();
+			}
+		}
 		public static string GetRelativePath(string filePath, string folder)
 		{
 			Uri pathUri = new Uri(filePath);
@@ -68,27 +83,40 @@ namespace Extract
 		public static string NormalizePath(string path)
 		{
 			return Path.GetFullPath(new Uri(path).LocalPath)
-					   .TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
-					   .ToUpperInvariant();
+						.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
+						.ToUpperInvariant();
 		}
-		public static ISerializedFile FindFile(FileCollection fileCollection, string path)
+		public static ISerializedFile FindFile(GameCollection fileCollection, string path)
 		{
-			return fileCollection.Files.First(f => f is SerializedFile sf && NormalizePath(sf.FilePath) == NormalizePath(path));
+			return fileCollection.GameFiles.Values.First(f => f is SerializedFile sf && NormalizePath(sf.FilePath) == NormalizePath(path));
+		}
+		public static IEnumerable<SerializedFile> GetSerializedFiles(GameCollection fileCollection)
+		{
+			return fileCollection.GameFiles.Values;
 		}
 		public static void ReplaceInFile(string filePath, string source, string replacement)
 		{
-			//if (!File.Exists(filePath)) return;
+			if (!File.Exists(filePath))
+			{
+				Logger.Log(LogType.Warning, LogCategory.Export, $"Could not perform line replace on {filePath}, file does not exist");
+				return;
+			}
 			var text = File.ReadAllText(filePath);
 			text = text.Replace(source, replacement);
 			File.WriteAllText(filePath, text);
 		}
 		public static void InsertInFile(string filePath, int index, string replacement)
 		{
-			//if (!File.Exists(filePath)) return;
+			if (!File.Exists(filePath))
+			{
+				Logger.Log(LogType.Warning, LogCategory.Export, $"Could not perform line insert on {filePath}, file does not exist");
+				return;
+			}
 			var lines = File.ReadAllLines(filePath).ToList();
 			lines.Insert(index, replacement);
 			File.WriteAllLines(filePath, lines);
 		}
+
 		public static List<string> GetManifestDependencies(string filePath)
 		{
 			var yaml = new YamlStream();
@@ -108,11 +136,12 @@ namespace Extract
 		public static T GetMember<T>(object obj, string name)
 		{
 			var field = obj.GetType().GetField(name, Util.AllBindingFlags);
-			if (field != null) {
+			if (field != null)
+			{
 				return (T)field.GetValue(obj);
 			}
 			var prop = obj.GetType().GetProperty(name, Util.AllBindingFlags);
-			if(prop != null)
+			if (prop != null)
 			{
 				return (T)prop.GetValue(obj);
 			}
@@ -185,9 +214,9 @@ namespace Extract
 			return sb.ToString();
 		}
 
-		public static void FixShaderBundle(FileCollection fileCollection)
+		public static void FixShaderBundle(GameCollection fileCollection)
 		{
-			var shaderBundle = fileCollection.Files.FirstOrDefault(f => f is SerializedFile sf && Path.GetFileName(sf.FilePath) == "shaders");
+			var shaderBundle = fileCollection.GameFiles.Values.FirstOrDefault(f => f is SerializedFile sf && Path.GetFileName(sf.FilePath) == "shaders");
 			if (shaderBundle != null)
 			{
 				foreach (var asset in shaderBundle.FetchAssets())
@@ -199,7 +228,7 @@ namespace Extract
 						using (MD5 md5 = MD5.Create())
 						{
 							byte[] md5Hash = md5.ComputeHash(Encoding.ASCII.GetBytes(shader.ValidName));
-							assetInfo.GUID = new EngineGUID(md5Hash);
+							assetInfo.GUID = new GUID(md5Hash);
 						}
 					}
 				}
@@ -207,16 +236,16 @@ namespace Extract
 		}
 		public static string GetName(uTinyRipper.Classes.Object asset)
 		{
-			if(asset is NamedObject no)
+			if (asset is NamedObject no)
 			{
 				return no.ValidName;
 			}
-			if(asset is MonoBehaviour mb && mb.IsScriptableObject)
+			if (asset is MonoBehaviour mb && mb.IsScriptableObject)
 			{
 				return mb.Name;
 			}
 			var nameProp = asset.GetType().GetProperty("Name");
-			if(nameProp != null)
+			if (nameProp != null)
 			{
 				return (string)nameProp.GetValue(asset);
 			}
@@ -226,12 +255,69 @@ namespace Extract
 		{
 			foreach (var asset in assets)
 			{
-				asset.AssetInfo.GUID = new EngineGUID(Guid.NewGuid());
+				asset.AssetInfo.GUID = new GUID(Guid.NewGuid());
 			}
 		}
 		public static void SetGUID(uTinyRipper.Classes.Object asset, Guid guid)
 		{
-			asset.AssetInfo.GUID = new EngineGUID(guid);
+			asset.AssetInfo.GUID = new GUID(guid);
+		}
+		static T CreateInstance<T>(params object[] parameters)
+		{
+			var instance = typeof(T)
+				.GetConstructors(AllBindingFlags)
+				.Single(c => c.GetParameters().Length == parameters.Length)
+				.Invoke(parameters);
+			return (T)instance;
+		}
+		public static GameCollection CreateGameCollection()
+		{
+			var layoutInfo = new LayoutInfo(new Version(), Platform.StandaloneWin64Player, TransferInstructionFlags.NoTransferInstructionFlags);
+			var layout = new AssetLayout(layoutInfo);
+			var parameters = new GameCollection.Parameters(layout);
+			parameters.ScriptBackend = ScriptingBackend.Mono;
+			var gameCollection = new GameCollection(parameters);
+			return gameCollection;
+		}
+		public static object LoadFile(string filepath)
+		{
+			var scheme = GameCollection.LoadScheme(filepath, Path.GetFileName(filepath));
+			object file = null;
+			if (scheme is SerializedFileScheme serializedFileScheme)
+			{
+				var platform = serializedFileScheme.Metadata != null &&
+					serializedFileScheme.Metadata.Hierarchy.Platform != 0 ?
+					serializedFileScheme.Metadata.Hierarchy.Platform
+					: Platform.StandaloneWin64Player;
+				var version = serializedFileScheme.Metadata != null ?
+					serializedFileScheme.Metadata.Hierarchy.Version
+					: new Version();
+				var layoutInfo = new LayoutInfo(version, platform, serializedFileScheme.Flags);
+				var layout = new AssetLayout(layoutInfo);
+				var parameters = new GameCollection.Parameters(layout);
+				var collection = new GameCollection(parameters);
+				file = Util.CreateInstance<SerializedFile>(collection, scheme);
+				typeof(SerializedFile).GetMethod("ReadData", AllBindingFlags)
+					.Invoke(file, new object[] { serializedFileScheme.Stream });
+			}
+			if (scheme is BundleFileScheme bundleFileScheme)
+			{
+				file = Util.CreateInstance<BundleFile>(scheme);
+			}
+			if (scheme is ArchiveFileScheme archiveFileScheme)
+			{
+				file = Util.CreateInstance<ArchiveFile>(scheme);
+			}
+			if (scheme is WebFileScheme webFileScheme)
+			{
+				file = Util.CreateInstance<WebFile>(scheme);
+			}
+			if (scheme is ResourceFileScheme resourceFileScheme)
+			{
+				file = Util.CreateInstance<ResourceFile>(scheme);
+			}
+			scheme.Dispose();
+			return file;
 		}
 	}
 }
